styles.css
The style sheet for the HTML code, this centering the image was way harder than it should've been. I started with another image which fit the vibe of this dumb tribute but I had to crop and I wanted to see if I could crop in CSS. What a pain that was trying to get the CSS code to work with a background for the image. Guessing it's best practice to crop your image outside of CSS first.